backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
pandas
WeasyPrint
gunicorn
python-dotenv

python-dotenv


